chore: fixed some styles for mobile + re-organized some files around
