+++ index.html
@@ -23,26 +23,29 @@
   </head>
   <body>
     <header>
-      <div class="logo">
-        <img
-          id="desktop-logo"
-          src="src/assets/images/Digital-NEST-logo.png"
-          alt="Digital NEST logo"
-        />
-        <img
-          id="mobile-logo"
-          src="src/assets/images/DN-logo.png"
-          alt="Digital NEST logo"
-        />
+      <div class="header-top">
+        <div class="logo">
+          <img
+            id="desktop-logo"
+            src="src/assets/images/Digital-NEST-logo.png"
+            alt="Digital NEST logo"
+          />
+          <img
+            id="mobile-logo"
+            src="src/assets/images/DN-logo.png"
+            alt="Digital NEST logo"
+          />
+        </div>
+        <h1 class="name">Nest Meet</h1>
+        <div class="header-spacer"></div>
       </div>
-      <h1 class="name">Nest Meet</h1>
-      <div class="header-spacer"></div>
+      <div class="header-bottom">
+        <div class="buttons-filters">
+          <div class="filters-container" id="filtersContainer"></div>
+        </div>
+      </div>
     </header>
     <main>
-      <div class="buttons-filters">
-        <div class="filters-container" id="filtersContainer"></div>
-      </div>
-
       <section class="week-card">
         <!-- TODO: Dynamically display the card info (interns, departments, etc..)-->
         <div class="inner-week-card">
@@ -134,4 +137,3 @@
     </footer>
   </body>
 </html>
-dropdowns)
