+++ package.json
@@ -5,9 +5,9 @@
   "main": "app.js",
   "scripts": {
     "lint:eslint": "eslint --config eslint.config.mjs . --fix",
-    "lint:style": "stylelint '**/*.css' --fix",
-    "lint:html": "htmlhint '**/*.html'",
-    "format": "prettier --write '**/*.{js,css,html}'",
+    "lint:style": "stylelint \"**/*.css\" --fix",
+    "lint:html": "htmlhint \"**/*.html\"",
+    "format": "prettier --write \"**/*.{js,css,html}\"",
     "check": "npm run format && npm run lint:eslint && npm run lint:style && npm run lint:html"
   },
   "keywords": [],
