+++ src/style.css
@@ -43,6 +43,15 @@ footer img {
 }
 
 header {
+  top: 0;
+  position: sticky;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  z-index: 200;
+}
+
+.header-top {
   display: flex;
   align-items: center;
   justify-content: space-between;
@@ -245,7 +254,6 @@ i#toggle-icon.rotate {
 
 /* Filter and Generate Buttons */
 .buttons-filters {
-  position: sticky;
   background-color: #ededed;
   top: 0;
   display: flex;
@@ -255,7 +263,8 @@ i#toggle-icon.rotate {
   width: 100%;
   padding: 20px 0;
   white-space: nowrap;
-  z-index: 10;
+  z-index: 999;
+  overflow-x: auto;
 }
 
 #schedule-button {
@@ -279,18 +288,10 @@ i#toggle-icon.rotate {
   font-size: medium;
 }
 
-#schedule-button {
-  background-color: #bebde3;
-}
-
-#clear-button {
-  background-color: #e3bdbd;
-}
-
 .filters-container {
   display: flex;
   gap: 20px;
-  overflow-x: visible; /* No scrolling on larger screens */
+  position: static;
 }
 
 .filter-container {
@@ -315,14 +316,17 @@ i#toggle-icon.rotate {
 .modal {
   display: none;
   position: absolute;
-  top: calc(100% + 5px);
-  left: 0;
-  width: 250px;
+}
+
+.fixed-modal-content {
+  position: fixed;
   background-color: white;
+  top: 170px;
+  width: 250px;
   border: 1px solid #ccc;
   box-shadow: 0 2px 10px rgb(0 0 0 / 10%);
   border-radius: 10px;
-  z-index: 5;
+  z-index: 9999;
 }
 
 .modal-content {
@@ -337,7 +341,6 @@ i#toggle-icon.rotate {
   width: 20px;
   height: 20px;
   background-color: transparent;
-  color: red;
   border: none;
   border-radius: 50%;
   display: flex;
@@ -345,11 +348,6 @@ i#toggle-icon.rotate {
   align-items: center;
   cursor: pointer;
   font-size: 18px;
-}
-
-.close-modal-button:hover {
-  background: none;
-  color: red;
 }
 
 .option-list {
@@ -446,26 +444,14 @@ i#toggle-icon.rotate {
   }
 
   .buttons-filters button {
-    padding: 1px;
-    font-size: small;
-    width: 150px;
+    padding: 10px;
   }
 
   .dropdown {
     font-size: small;
   }
-}
 
-/* Tablet and mobile view (up to 1024px wide) */
-@media screen and (max-width: 950px) {
-  .filter-container {
-    width: 100%;
-    overflow-x: auto;
-    max-width: 170px;
-  }
-
-  .modal {
-    position: fixed;
+  .fixed-modal-content {
     top: 50%;
     left: 50%;
     transform: translate(-50%, -50%);
@@ -473,6 +459,21 @@ i#toggle-icon.rotate {
     max-width: 300px;
     max-height: 80vh;
     overflow-y: auto;
+  }
+
+  .filter-container,
+  .filter-button,
+  #schedule-button,
+  #clear-filters-button {
+    width: 150px !important;
+    font-size: 16px;
+  }
+}
+
+@media screen and (width <= 950px) {
+  .filter-container {
+    width: 100%;
+    max-width: 170px;
   }
 
   .option-item {
